PRF_PROG1_20232_HernandezJuanPablo_EspitiaAlejandra_GuerreroValentina_PenicheGianfranco/src/co/edu/unbosque/test/ReporteTest.java
package co.edu.unbosque.test;

import static org.junit.Assert.assertEquals;
import org.junit.Before;
import org.junit.Test;

import co.edu.unbosque.model.Reporte;

/**
 * Clase de pruebas para la clase Reporte.
 */
public class ReporteTest {

    private Reporte reporte;

    @Before
    public void setUp() {
        reporte = new Reporte();
    }

    /**
     * Prueba para verificar la generación correcta del reporte de lista de clientes.
     */
    @Test
    public void testGenerarReporteListaClientes() {
        String resultado = reporte.generarReporteListaClientes();
        // Aquí puedes verificar el resultado esperado o realizar otras comprobaciones.
        // assertEquals("Resultado Esperado", resultado);
    }

    /**
     * Prueba para verificar la generación correcta del reporte de valor total de apuestas.
     */
    @Test
    public void testGenerarReporteValorTotalApuestas() {
        String resultado = reporte.generarReporteValorTotalApuestas();
        // Aquí puedes verificar el resultado esperado o realizar otras comprobaciones.
        // assertEquals("Resultado Esperado", resultado);
    }

    /**
     * Prueba para verificar la generación correcta del reporte de detalles de apuestas por sede.
     */
    @Test
    public void testGenerarReporteDetallesApuestasPorSede() {
        String resultado = reporte.generarReporteDetallesApuestasPorSede();
        // Aquí puedes verificar el resultado esperado o realizar otras comprobaciones.
        // assertEquals("Resultado Esperado", resultado);
    }

    /**
     * Prueba para verificar la generación correcta del reporte de detalles de apuestas por apostador.
     */
    @Test
    public void testGenerarReporteDetallesApuestasPorApostador() {
        String resultado = reporte.generarReporteDetallesApuestasPorApostador();
        // Aquí puedes verificar el resultado esperado o realizar otras comprobaciones.
        // assertEquals("Resultado Esperado", resultado);
    }

    /**
     * Prueba para verificar la generación correcta del reporte de valor total de apuestas por sede.
     */
    @Test
    public void testGenerarReporteValorTotalApuestasPorSede() {
        String resultado = reporte.generarReporteValorTotalApuestasPorSede();
        // Aquí puedes verificar el resultado esperado o realizar otras comprobaciones.
        // assertEquals("Resultado Esperado", resultado);
    }
}
